move the dataavailabilityheader from types to da
