+++ pkg/da/data_availability_header_test.go
@@ -1,0 +1,17 @@
+package da
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+)
+
+func TestNilDataAvailabilityHeaderHashDoesntCrash(t *testing.T) {
+	// This follows RFC-6962, i.e. `echo -n '' | sha256sum`
+	var emptyBytes = []byte{0xe3, 0xb0, 0xc4, 0x42, 0x98, 0xfc, 0x1c, 0x14, 0x9a, 0xfb, 0xf4, 0xc8,
+		0x99, 0x6f, 0xb9, 0x24, 0x27, 0xae, 0x41, 0xe4, 0x64, 0x9b, 0x93, 0x4c, 0xa4, 0x95, 0x99, 0x1b,
+		0x78, 0x52, 0xb8, 0x55}
+
+	assert.Equal(t, emptyBytes, (*DataAvailabilityHeader)(nil).Hash())
+	assert.Equal(t, emptyBytes, new(DataAvailabilityHeader).Hash())
+}
