+++ proto/tendermint/types/block.proto
@@ -5,10 +5,11 @@ option go_package = "github.com/celestiaorg/celestia-core/proto/tendermint/types
 
 import "gogoproto/gogo.proto";
 import "tendermint/types/types.proto";
+import "tendermint/da/data_availability_header.proto";
 
 message Block {
   Header                 header                   = 1 [(gogoproto.nullable) = false];
   Data                   data                     = 2 [(gogoproto.nullable) = false];
-  DataAvailabilityHeader data_availability_header = 3;
+  tendermint.da.DataAvailabilityHeader data_availability_header = 3;
   Commit                 last_commit              = 4;
 }
